Replace subreddit search with a dialog
Also allows custom subreddits not showing up in the search to be added.
Looks them up via a call to the About API end-point. If valid data is
returned then it will be added to the database and thus the sidebar.
Stripped out search of the Sidebar and made it just filter only.